inget viktigt

+++ WinApp.cpp
@@ -92,7 +92,7 @@ int WinApp::run(void)
 			m_timer.tick();
 
 			// Add update- and render-stuff here!
-			m_game.Update(m_timer.getDeltaTime());
+			m_game.Update((float)m_timer.getDeltaTime());
 			m_game.Draw();
 		}
 	}
